small rng overhaul
makes it so game takes in a default_random_engine for a parameter instead of a seed, because reseeding the rng makes everything act funky i think. Having it be a parameter of Game is good so every method can use it. enemy generation is working well also i think there are no segmentation faults